# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/edit/{email}": {
      "patch": {
        "operationId": "UserController_editUserBySuperAdmin",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/event": {
      "post": {
        "operationId": "EventEditionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventEditionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EventEdition"
        ]
      },
      "get": {
        "operationId": "EventEditionController_getAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventEdition"
        ]
      }
    },
    "/event/create-from-event-edition-form": {
      "post": {
        "operationId": "EventEditionController_createFromEventEditionForm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFromEventEditionFormDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EventEdition"
        ]
      }
    },
    "/event/update-from-event-edition-form/{id}": {
      "put": {
        "operationId": "EventEditionController_updateFromEventEditionForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFromEventEditionFormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventEdition"
        ]
      }
    },
    "/event/{id}": {
      "get": {
        "operationId": "EventEditionController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventEdition"
        ]
      },
      "put": {
        "operationId": "EventEditionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventEditionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EventEdition"
        ]
      },
      "delete": {
        "operationId": "EventEditionController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EventEdition"
        ]
      }
    },
    "/event/year/{year}": {
      "get": {
        "operationId": "EventEditionController_getByYear",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventEdition"
        ]
      }
    },
    "/event/active/{id}": {
      "patch": {
        "operationId": "EventEditionController_setActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EventEdition"
        ]
      }
    },
...
    "/presentation-block/event-edition/{eventEditionId}": {
      "get": {
        "operationId": "PresentationBlockController_findAllByEventEditionId",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      }
    },
...
    "/evaluations": {
      "put": {
        "operationId": "EvaluationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of evaluations (one for each criteria), all for the same submissionId.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateEvaluationDto"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The evaluations have been created successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new evaluations.",
        "tags": [
          "Evaluation"
        ]
      },
      "get": {
        "operationId": "EvaluationController_find",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter evaluations by user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return evaluations."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find evaluations",
        "tags": [
          "Evaluation"
        ]
      }
    },
    "/evaluations/user": {
      "get": {
        "operationId": "EvaluationController_findByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find evaluations by user",
        "tags": [
          "Evaluation"
        ]
      }
    },
    "/evaluations/submission/{submissionId}/final-grade": {
      "get": {
        "operationId": "EvaluationController_calculateFinalGrade",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Evaluation"
        ]
      }
    },
...
    "/room/event-edition/{eventEditionId}": {
      "get": {
        "operationId": "RoomController_findAll",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
...
    "/certificate/event-edition/{eventEditionId}": {
      "get": {
        "operationId": "CertificateController_downloadCertificate",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Certificate"
        ]
      }
    }
  },
  "info": {
    "title": "WEPGCOMP API",
    "description": "Portal WEPGCOMP - API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateEventEditionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFromEventEditionFormDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateFromEventEditionFormDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEventEditionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateEvaluationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}